tabs: Remove reliance on global text tags
Gtk text tags are associated with a specific TextBuffer object. Since we are
going to have multiple dynamically created TextBuffer objects we need to have
a set of tags for each one. So we create a function to generate the tags and
associate them with a given TextBuffer. Then we start looking up tags by name
rather than using global stored references.